move ValidationSummary to common
